Emit linear num of instr for Extract and Create.
Previously O(n^2) instructions were emitted to compute addresses of
nested pair fields, which typically should be constant folded by an
optimizer.

This change does not seem to have a major effect on generation time on
current tests, but, in the case of exceptionally large products, making
the number of instructions linear could have a significant effect.