Rethink memoization one more time.
The idea now is to statically map results to the given rules class type.
This should allow independent program modules to use Infers without need
for global coordination.

It seems to be an anti-pattern to define rules for "unsupported" cases
and it is better to let rules fail in case of unsupported types.  Rules
classes need to be designed with potential extensibility in mind.  Users
can extend existing rules by defining a local rules class that depends
on the base rules and defines additional rules.