Implicit rules for constructors seem hazardous.
Constructor rules were originally added to support dependencies to other
rules, but it is no longer required as dependencies are specified via
attributes.

The problem with constructors is that, when fully supported, they tend
to allow drawing values out of thin air, which defeats the idea of
controlling which values are generated via carefully defined rules.

It is likely better to use some sort of explicit generated rules for
accessing constructors when desired.